+++ tests/test_language.py
@@ -105,12 +105,15 @@ def test_invalid_inputs():
 
 
 def test_data_last_updated():
-    assert DATA_LAST_UPDATED == datetime.date(2025, 1, 15), "Need to update README.md"
+    assert DATA_LAST_UPDATED == datetime.date(2025, 10, 15), "Need to update README.md"
 
 
 def test_all_languages():
     assert type(ALL_LANGUAGES) is set
-    assert len(ALL_LANGUAGES) == 8307, "Need to update README.md"
+    # Defining `len_all_languages` so that when the assertion fails,
+    # the error message is more informative.
+    len_all_languages = len(ALL_LANGUAGES)
+    assert len_all_languages == 8311, "Need to update README.md"
     lang = list(ALL_LANGUAGES)[0]
     assert type(lang) is Language